Fix border radius
The definition of the `border-radius` property for the primary home page buttons was overridden by default Bootstrap styles. It takes precedence if it's placed within the `.group-main .btn` declarations.

After making this change, the `font` property is left alone. By moving it along with `border-radius`, the `.btn-huge` declaration is no longer needed and can be removed completely.