Venues, initially, finally
I am finally committing venues. It's actually a good thing that I didn't do it sooner because the methodology behind retrieving them has changed quite a bit.

When a user logs in, in addition to retrieving their groups, the venues associated with those groups are now collected as well. As part of the process, the `lon` and `lat` fields are stripped and placed into a `loc` field, an array with can be used for 2d geospatial queries right in Mongo.