Ensure all the indexes
Okay, so right now there's just one index, but the `ensure_index()` method  will run whenever the server starts up--well, before it processes its first request anyway--and make sure that all of the desired Mongo indexes are in place.

This is the 2d index on the `loc` field of the `venues` collection.