Add capacity field to venue form
One of the most important things when trying to pick a venue is knowing how many people can fit in the space. The field is now a part a `VenueEditForm` (renamed from `VenueClaimForm` in this revision).

This revision also makes an attempt at making a distinction between claiming a venue and editing one that has already been claimed. The `confirm` field is no longer a part of `VenueEditForm`. Instead, a new `VenueClaimForm`--a subclass of `VenueEditForm`--is being added and given this one additional field. The view will select the correct form to use based on whether or not a venue has already been claimed. The same logic is applied in the template as well, also altering the submit button text.

The final change being added here is more related to the `VenueEditForm`. As originally written, a `User` was supplied to the form for default values. Unfortunately the way this was written, the user's values were always used when editing the venue. Here I'm refactoring the view to get the code from the venue first, then fall back on the values from the user.

I'm not totally convinced that this is the best way to do some of these things, but I'd like to get these changes saved and come back to them later.

This fixes #15.