Remove default capacity value
I thought I had removed the default value of capacity being set in `VenueEditForm`, but I guess not.

The other half of this revision is a revert of 52e06eb. Since we aren't currently working with unclaimed venues anywhere that would utilize the capacity field, there is no need to set it on a document when it is first created.