Add venue questionnaire
As part of claiming/editing a venue, users can now enter the following optional information:
* if the venue needs a list of names
* if the venue can sponsor food
* if the venue has A/V equipment
* if the venue has enough seating
* special instructions for getting to the venue

To help with this change, the `Venue.claim()` method has been refactored. It now sets the values to its attributes and calls its `save()` method.

Oh, and the claim form has been structured a little better, but it still needs a designer to show it some love.